% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_api.R
\name{ds_authn}
\alias{ds_authn}
\alias{ds_login}
\alias{ds_logout}
\alias{ds_status}
\title{DSpace authn functions}
\usage{
ds_login(x = NULL, ..., simplifyVector = TRUE)

ds_logout(x = NULL, ..., simplifyVector = TRUE)

ds_status(x = NULL, ..., simplifyVector = TRUE)
}
\description{
DSpace authn functions
}
\details{
\url{https://github.com/DSpace/RestContract/blob/main/authentication.md}

\url{https://github.com/DSpace/RestContract/blob/main/authentication.md}

\url{https://github.com/DSpace/RestContract/blob/main/authentication.md}
}
