% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_api.R
\name{ds_config}
\alias{ds_config}
\alias{ds_harvestermetadata}
\alias{ds_submissionaccessoptions}
\alias{ds_submissiondefinitions}
\alias{ds_submissionforms}
\alias{ds_submissionsections}
\alias{ds_submissionuploads}
\alias{ds_workflowdefinitions}
\alias{ds_workflowsteps}
\title{DSpace config functions}
\usage{
ds_harvestermetadata(x = NULL, ..., simplifyVector = TRUE)

ds_submissionaccessoptions(x = NULL, ..., simplifyVector = TRUE)

ds_submissiondefinitions(x = NULL, ..., simplifyVector = TRUE)

ds_submissionforms(x = NULL, ..., simplifyVector = TRUE)

ds_submissionsections(x = NULL, ..., simplifyVector = TRUE)

ds_submissionuploads(x = NULL, ..., simplifyVector = TRUE)

ds_workflowdefinitions(x = NULL, ..., simplifyVector = TRUE)

ds_workflowsteps(x = NULL, ..., simplifyVector = TRUE)
}
\description{
DSpace config functions
}
\details{
\url{https://github.com/DSpace/RestContract/blob/main/harvestermetadata.md}

\url{https://github.com/DSpace/RestContract/blob/main/submissionaccessoptions.md}

\url{https://github.com/DSpace/RestContract/blob/main/submissiondefinitions.md}

\url{https://github.com/DSpace/RestContract/blob/main/submissionforms.md}

\url{https://github.com/DSpace/RestContract/blob/main/submissionsections.md}

\url{https://github.com/DSpace/RestContract/blob/main/submissionuploads.md}

\url{https://github.com/DSpace/RestContract/blob/main/workflowdefinitions.md}

\url{https://github.com/DSpace/RestContract/blob/main/workflowsteps.md}
}
