% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_api.R
\name{ds_core}
\alias{ds_core}
\alias{ds_bitstreamformats}
\alias{ds_bitstreams}
\alias{ds_bundles}
\alias{ds_collections}
\alias{ds_communities}
\alias{ds_epersons}
\alias{ds_groups}
\alias{ds_items}
\alias{ds_itemtemplates}
\alias{ds_metadatafields}
\alias{ds_metadataschemas}
\title{DSpace core functions}
\usage{
ds_bitstreamformats(x = NULL, ..., simplifyVector = TRUE)

ds_bitstreams(x = NULL, ..., simplifyVector = TRUE)

ds_bundles(x = NULL, ..., simplifyVector = TRUE)

ds_collections(x = NULL, ..., simplifyVector = TRUE)

ds_communities(x = NULL, ..., simplifyVector = TRUE)

ds_epersons(x = NULL, ..., simplifyVector = TRUE)

ds_groups(x = NULL, ..., simplifyVector = TRUE)

ds_items(x = NULL, ..., simplifyVector = TRUE)

ds_itemtemplates(x = NULL, ..., simplifyVector = TRUE)

ds_metadatafields(x = NULL, ..., simplifyVector = TRUE)

ds_metadataschemas(x = NULL, ..., simplifyVector = TRUE)
}
\description{
DSpace core functions
}
\details{
\url{https://github.com/DSpace/RestContract/blob/main/bitstreamformats.md}

\url{https://github.com/DSpace/RestContract/blob/main/bitstreams.md}

\url{https://github.com/DSpace/RestContract/blob/main/bundles.md}

\url{https://github.com/DSpace/RestContract/blob/main/collections.md}

\url{https://github.com/DSpace/RestContract/blob/main/communities.md}

\url{https://github.com/DSpace/RestContract/blob/main/epersons.md}

\url{https://github.com/DSpace/RestContract/blob/main/epersongroups.md}

\url{https://github.com/DSpace/RestContract/blob/main/items.md}

\url{https://github.com/DSpace/RestContract/blob/main/itemtemplates.md}

\url{https://github.com/DSpace/RestContract/blob/main/metadatafields.md}

\url{https://github.com/DSpace/RestContract/blob/main/metadataschemas.md}
}
