% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_api.R
\name{ds_discover}
\alias{ds_discover}
\alias{ds_browses}
\alias{ds_search}
\title{DSpace discover functions}
\usage{
ds_browses(x = NULL, ..., simplifyVector = TRUE)

ds_search(x = NULL, ..., simplifyVector = TRUE)
}
\description{
DSpace discover functions
}
\details{
\url{https://github.com/DSpace/RestContract/blob/main/browses.md}

\url{https://github.com/DSpace/RestContract/blob/main/search-endpoint.md}
}
