% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_api.R
\name{ds_eperson}
\alias{ds_eperson}
\alias{ds_orcidhistories}
\alias{ds_orcidqueues}
\alias{ds_profiles}
\title{DSpace eperson functions}
\usage{
ds_orcidhistories(x = NULL, ..., simplifyVector = TRUE)

ds_orcidqueues(x = NULL, ..., simplifyVector = TRUE)

ds_profiles(x = NULL, ..., simplifyVector = TRUE)
}
\description{
DSpace eperson functions
}
\details{
\url{https://github.com/DSpace/RestContract/blob/main/orcidhistories.md}

\url{https://github.com/DSpace/RestContract/blob/main/orcidqueues.md}

\url{https://github.com/DSpace/RestContract/blob/main/profiles.md}
}
