% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_api.R
\name{ds_integration}
\alias{ds_integration}
\alias{ds_qualityassuranceevents}
\alias{ds_qualityassurancesources}
\alias{ds_qualityassurancetopics}
\alias{ds_suggestions}
\alias{ds_suggestionsources}
\alias{ds_suggestiontargets}
\title{DSpace integration functions}
\usage{
ds_qualityassuranceevents(x = NULL, ..., simplifyVector = TRUE)

ds_qualityassurancesources(x = NULL, ..., simplifyVector = TRUE)

ds_qualityassurancetopics(x = NULL, ..., simplifyVector = TRUE)

ds_suggestions(x = NULL, ..., simplifyVector = TRUE)

ds_suggestionsources(x = NULL, ..., simplifyVector = TRUE)

ds_suggestiontargets(x = NULL, ..., simplifyVector = TRUE)
}
\description{
DSpace integration functions
}
\details{
\url{https://github.com/DSpace/RestContract/blob/main/qualityassuranceevents.md}

\url{https://github.com/DSpace/RestContract/blob/main/qualityassurancesources.md}

\url{https://github.com/DSpace/RestContract/blob/main/qualityassurancetopics.md}

\url{https://github.com/DSpace/RestContract/blob/main/suggestions.md}

\url{https://github.com/DSpace/RestContract/blob/main/suggestionsources.md}

\url{https://github.com/DSpace/RestContract/blob/main/suggestiontargets.md}
}
