% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.R
\name{ds_collections}
\alias{ds_collections}
\alias{ds_items}
\alias{ds_communities}
\title{List DSpace Collections, Items, or Communities}
\usage{
ds_collections(x = NULL, ..., simplifyVector = TRUE)

ds_items(x = NULL, ..., simplifyVector = TRUE)

ds_communities(x = NULL, ..., simplifyVector = TRUE)
}
\description{
List DSpace Collections, Items, or Communities
}
