% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_api.R
\name{ds_submission}
\alias{ds_submission}
\alias{ds_duplicates}
\alias{ds_vocabularies}
\alias{ds_vocabularyEntryDetails}
\alias{ds_workspaceitems}
\title{DSpace submission functions}
\usage{
ds_duplicates(x = NULL, ..., simplifyVector = TRUE)

ds_vocabularies(x = NULL, ..., simplifyVector = TRUE)

ds_vocabularyEntryDetails(x = NULL, ..., simplifyVector = TRUE)

ds_workspaceitems(x = NULL, ..., simplifyVector = TRUE)
}
\description{
DSpace submission functions
}
\details{
\url{https://github.com/DSpace/RestContract/blob/main/duplicates.md}

\url{https://github.com/DSpace/RestContract/blob/main/vocabularies.md}

\url{https://github.com/DSpace/RestContract/blob/main/vocabularyEntryDetails.md}

\url{https://github.com/DSpace/RestContract/blob/main/workspaceitems.md}
}
