% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_api.R
\name{ds_system}
\alias{ds_system}
\alias{ds_systemwidealerts}
\title{DSpace system functions}
\usage{
ds_systemwidealerts(x = NULL, ..., simplifyVector = TRUE)
}
\description{
DSpace system functions
}
\details{
\url{https://github.com/DSpace/RestContract/blob/main/systemwidealerts.md}
}
