% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_api.R
\name{ds_tools}
\alias{ds_tools}
\alias{ds_feedbacks}
\title{DSpace tools functions}
\usage{
ds_feedbacks(x = NULL, ..., simplifyVector = TRUE)
}
\description{
DSpace tools functions
}
\details{
\url{https://github.com/DSpace/RestContract/blob/main/feedbacks.md}
}
