% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_api.R
\name{ds_versioning}
\alias{ds_versioning}
\alias{ds_versionhistories}
\alias{ds_versions}
\title{DSpace versioning functions}
\usage{
ds_versionhistories(x = NULL, ..., simplifyVector = TRUE)

ds_versions(x = NULL, ..., simplifyVector = TRUE)
}
\description{
DSpace versioning functions
}
\details{
\url{https://github.com/DSpace/RestContract/blob/main/versionhistories.md}

\url{https://github.com/DSpace/RestContract/blob/main/versions.md}
}
