% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_api.R
\name{ds_workflow}
\alias{ds_workflow}
\alias{ds_claimedtasks}
\alias{ds_pooltasks}
\alias{ds_workflowitems}
\title{DSpace workflow functions}
\usage{
ds_claimedtasks(x = NULL, ..., simplifyVector = TRUE)

ds_pooltasks(x = NULL, ..., simplifyVector = TRUE)

ds_workflowitems(x = NULL, ..., simplifyVector = TRUE)
}
\description{
DSpace workflow functions
}
\details{
\url{https://github.com/DSpace/RestContract/blob/main/claimedtasks.md}

\url{https://github.com/DSpace/RestContract/blob/main/pooltasks.md}

\url{https://github.com/DSpace/RestContract/blob/main/workflowitems.md}
}
