% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req.R
\name{dspace_request}
\alias{dspace_request}
\title{Wrapper function for \code{httr2::request}}
\usage{
dspace_request(base_url = getOption("dpacer.base_url"), ...)
}
\description{
Wrapper function for \code{httr2::request}
}
