% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dspacer-package.R
\docType{package}
\name{dspacer-package}
\alias{dspacer}
\alias{dspacer-package}
\title{dspacer: Access the DSpace REST API}
\description{
Supports read acess to the DSpace REST API.
}
\author{
\strong{Maintainer}: Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [copyright holder]

}
\keyword{internal}
